added dao, controller and service for speciality

+++ src/main/java/com/stefanini/appointmentapp/service/impl/DepartmentServiceImpl.java
@@ -5,6 +5,7 @@ import com.stefanini.appointmentapp.entities.Department;
 import com.stefanini.appointmentapp.service.DepartmentService;
 import org.springframework.stereotype.Service;
 
+import javax.persistence.NoResultException;
 import java.util.Set;
 
 
@@ -34,11 +35,16 @@ public class DepartmentServiceImpl implements DepartmentService {
     @Override
     public void delete(long departmentId) {
         Department departmentToDelete = departmentDAO.findById(departmentId);
+
+        if (departmentToDelete == null) {
+            throw new NoResultException("Department delete: department with id = " + departmentId + " not found");
+        }
+
         departmentDAO.delete(departmentToDelete);
     }
 
     @Override
-    public Department update(Department department, long id) {
+    public Department update(Department department) {
         return departmentDAO.update(department);
     }
 }
